Revert "ROX-33958: BusyBox-style binary consolidation for main image (#19819)"
This reverts commit b54ecbf2f0c30a9f017243f22b9380bc997e5eb1.
