+++ central/main.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"syscall"
 	"time"
 
@@ -231,20 +230,18 @@ import (
 	"github.com/stackrox/rox/pkg/sync"
 	"github.com/stackrox/rox/pkg/utils"
 	pkgVersion "github.com/stackrox/rox/pkg/version"
-
-	// BusyBox-style consolidation - import app packages
-	complianceapp "github.com/stackrox/rox/compliance/cmd/compliance/app"
-	roxagentapp "github.com/stackrox/rox/compliance/virtualmachines/roxagent/app"
-	configcontrollerapp "github.com/stackrox/rox/config-controller/app"
-	migratorapp "github.com/stackrox/rox/migrator/app"
-	roxctlapp "github.com/stackrox/rox/roxctl/app"
-	admissioncontrolapp "github.com/stackrox/rox/sensor/admission-control/app"
-	kubernetessensorapp "github.com/stackrox/rox/sensor/kubernetes/app"
-	sensorupgraderapp "github.com/stackrox/rox/sensor/upgrader/app"
 )
 
 var (
 	log = logging.CreateLogger(logging.CurrentModule(), 0)
+
+	authProviderBackendFactories = map[string]authproviders.BackendFactoryCreator{
+		oidc.TypeName:                oidc.NewFactory,
+		"auth0":                      oidc.NewFactory, // legacy
+		saml.TypeName:                saml.NewFactory,
+		authProviderUserpki.TypeName: authProviderUserpki.NewFactoryFactory(tlsconfig.ManagerInstance()),
+		iap.TypeName:                 iap.NewFactory,
+	}
 
 	imageIntegrationContext = sac.WithGlobalAccessScopeChecker(context.Background(),
 		sac.AllowFixedScopes(
@@ -281,7 +278,7 @@ func runSafeMode() {
 	log.Info("Central terminated")
 }
 
-func centralRun() {
+func main() {
 	defer utils.IgnoreError(log.InnerLogger.Sync)
 
 	premain.StartMain()
@@ -560,17 +557,6 @@ func startGRPCServer() {
 
 	// Create the registry of applied auth providers.
 	registry := authProviderRegistry.Singleton()
-
-	// Initialize auth provider backend factories. This must be done here (not at package level)
-	// because tlsconfig.ManagerInstance() requires certificate files that may not be available
-	// for all entry points in the consolidated binary.
-	authProviderBackendFactories := map[string]authproviders.BackendFactoryCreator{
-		oidc.TypeName:                oidc.NewFactory,
-		"auth0":                      oidc.NewFactory, // legacy
-		saml.TypeName:                saml.NewFactory,
-		authProviderUserpki.TypeName: authProviderUserpki.NewFactoryFactory(tlsconfig.ManagerInstance()),
-		iap.TypeName:                 iap.NewFactory,
-	}
 
 	// env.EnableOpenShiftAuth signals the desire but does not guarantee Central
 	// is configured correctly to talk to the OpenShift's OAuth server. If this
@@ -1078,34 +1064,4 @@ func waitForTerminationSignal() {
 		osutils.Restart()
 	}
 	log.Info("Central terminated")
-}
-
-func main() {
-	// BusyBox-style dispatcher: check how we were called
-	binaryName := filepath.Base(os.Args[0])
-
-	switch binaryName {
-	case "central":
-		centralRun()
-	case "migrator":
-		migratorapp.Run()
-	case "compliance":
-		complianceapp.Run()
-	case "kubernetes-sensor":
-		kubernetessensorapp.Run()
-	case "sensor-upgrader":
-		sensorupgraderapp.Run()
-	case "admission-control":
-		admissioncontrolapp.Run()
-	case "config-controller":
-		configcontrollerapp.Run()
-	case "roxagent":
-		roxagentapp.Run()
-	case "roxctl":
-		roxctlapp.Run()
-	default:
-		// Default to central if called with unknown name
-		log.Warnf("Unknown binary name %q, defaulting to central mode", binaryName)
-		centralRun()
-	}
 }
