+++ sensor/admission-control/certs.go
@@ -1,4 +1,4 @@
-package app
+package main
 
 import (
 	"os"
